add meta page, refactor page header
