--- website/frontend/src/components/Meta/Meta.scss
@@ -1,15 +1,0 @@
-.meta-form {
-  position: fixed;
-  top: 0;
-  left: 0;
-  height: 100%;
-  width: 100vw;
-  overflow-y: scroll;
-  z-index: 100;
-}
-
-@media (max-width: 768px) {
-  .meta-form {
-    z-index: 100;
-  }
-}
