+++ website/frontend/src/components/blocks/PageHeader/PageHeader.tsx
@@ -1,47 +1,23 @@
-import React, { VFC } from 'react';
-import { PAGE_TYPES } from 'consts';
+import styles from './PageHeader.module.scss';
+
 import { fileNameHandler } from 'helpers';
-import ArrowBack from 'assets/images/arrow_back.svg';
-import ProgramIllustration from 'assets/images/program_icon.svg';
-import close from 'assets/images/close.svg';
-import { Wallet } from '../Wallet/Wallet';
-import './PageHeader.scss';
+import { BackButton } from 'components/BackButton/BackButton';
+import ProgramIllustrationSVG from 'assets/images/program_icon.svg';
 
 type Props = {
-  programName: string;
-  pageType: string;
-  handleClose: () => void;
+  title: string;
+  fileName: string;
 };
 
-export const PageHeader: VFC<Props> = ({ programName, pageType, handleClose }) => {
-  let headInscription = 'New request to';
-  let formId = 'message-form';
-  if (pageType === PAGE_TYPES.EDITOR_PAGE) {
-    headInscription = 'Edit code';
-  } else if (pageType === PAGE_TYPES.META_FORM_PAGE) {
-    headInscription = 'Upload metadata';
-    formId = 'meta-form';
-  }
-
-  return (
-    <div className="message-header">
-      <div className="message-header--info">
-        <button
-          type="button"
-          aria-label="arrowBack"
-          onClick={() => handleClose()}
-          className="message-header--info__back"
-        >
-          <img src={ArrowBack} alt="back" />
-        </button>
-        <h2 className="message-header--info__text">{headInscription}</h2>
-        <img src={ProgramIllustration} alt="program" className="message-header--info__icon" />
-        <h2 className="message-header--info__filename">{fileNameHandler(programName)}</h2>
+export const PageHeader = ({ title, fileName }: Props) => (
+  <div className={styles.pageHeader}>
+    <BackButton className={styles.headerBackBtn} />
+    <div className={styles.headerInfo}>
+      <h2 className={styles.infoTitle}>{title}</h2>
+      <div className={styles.infoFile}>
+        <img className={styles.fileIcon} src={ProgramIllustrationSVG} alt="file" />
+        <span className={styles.fileName}>{fileNameHandler(fileName)}</span>
       </div>
-      <Wallet />
-      <button type="reset" aria-label="closeButton" form={formId} className="message-header--info__close">
-        <img src={close} alt="close" />
-      </button>
     </div>
-  );
-};
+  </div>
+);
