+++ website/frontend/src/components/pages/Send/Send.scss
@@ -1,30 +1,5 @@
 @use 'assets/scss/variables' as *;
 
-.send-message__header {
-  display: flex;
-  align-content: center;
-  margin-bottom: 40px;
-
-  .send-message__button-back {
-    margin-right: 20px;
-    padding: 0;
-  }
-
-  .send-message__header-text {
-    font-size: 21px;
-    font-weight: 600;
-    text-transform: uppercase;
-    color: $textColorSecondary;
-    &_colored {
-      color: $textColor;
-    }
-  }
-
-  .send-message__header-icon {
-    margin: 0 10px 0 20px;
-  }
-}
-
 .send-message__block {
   background-color: $bkgActiveButtonColor;
   border-radius: 2px;
