+++ website/frontend/src/features/Editor/EditorPage/EditorPage.tsx
@@ -7,7 +7,7 @@ import { useNavigate } from 'react-router-dom';
 import clsx from 'clsx';
 import get from 'lodash.get';
 
-import { PageHeader } from 'components/blocks/PageHeader/PageHeader';
+import { PageHeader } from 'components/blocks/legacy/PageHeader/PageHeader';
 import { useEditor } from 'hooks';
 import { EDITOR_BTNS, PAGE_TYPES, WASM_COMPILER_BUILD, LOCAL_STORAGE } from 'consts';
